// NextJS/PortfolioApp/types/expEdu.ts
export type EduExp =
{
    profile: string;
    institution: string;
    coursePT: string;
    courseEN: string;
    courseES: string;
    courseFR: string;
    locationPT: string;
    locationEN: string;
    locationFR: string;
    locationES: string;
    detailsPT: string;
    detailsEN: string;
    detailsFR: string;
    detailsES: string;
    webSite:string;
    active:boolean;
    startedAt: string;
    endedAt: string;
}